+++ CHANGELOG.md
@@ -2,6 +2,57 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.15](https://github.com/salesforcecli/plugin-dev/compare/v0.0.14...v0.0.15) (2022-08-31)
+
+
+### Bug Fixes
+
+* add error message for invalid plugin name ([ea8d9a3](https://github.com/salesforcecli/plugin-dev/commit/ea8d9a35619b8b7e45945ef0f1a233ccababd7cf))
+
+### [0.0.14](https://github.com/salesforcecli/plugin-dev/compare/v0.0.13...v0.0.14) (2022-08-31)
+
+
+### Bug Fixes
+
+* update template ([4799859](https://github.com/salesforcecli/plugin-dev/commit/47998591d7be474b28a608200390accab70977d5))
+
+### [0.0.13](https://github.com/salesforcecli/plugin-dev/compare/v0.0.12...v0.0.13) (2022-08-31)
+
+
+### Bug Fixes
+
+* default github username ([565a908](https://github.com/salesforcecli/plugin-dev/commit/565a908b79e9025a72925b6ecf96f7590a01595a))
+
+### [0.0.12](https://github.com/salesforcecli/plugin-dev/compare/v0.0.11...v0.0.12) (2022-08-30)
+
+
+### Features
+
+* make plugin generator 3PP friendly ([#104](https://github.com/salesforcecli/plugin-dev/issues/104)) ([301a36b](https://github.com/salesforcecli/plugin-dev/commit/301a36b8780368b2a32d86eff22123f6661bfed2))
+
+### [0.0.11](https://github.com/salesforcecli/plugin-dev/compare/v0.0.10...v0.0.11) (2022-08-16)
+
+### [0.0.10](https://github.com/salesforcecli/plugin-dev/compare/v0.0.9...v0.0.10) (2022-08-10)
+
+
+### Bug Fixes
+
+* yeoman generator ([#98](https://github.com/salesforcecli/plugin-dev/issues/98)) ([acb2122](https://github.com/salesforcecli/plugin-dev/commit/acb2122322cc3d422c9067f9ba05f0fc2964e473))
+
+### [0.0.9](https://github.com/salesforcecli/plugin-dev/compare/v0.0.8...v0.0.9) (2022-08-03)
+
+
+### Bug Fixes
+
+* use https instead of ssh for git cloning ([f248be6](https://github.com/salesforcecli/plugin-dev/commit/f248be69f38705adcfff376e70d3bf13e58465c3))
+
+### [0.0.8](https://github.com/salesforcecli/plugin-dev/compare/v0.0.7...v0.0.8) (2022-07-27)
+
+
+### Bug Fixes
+
+* await generators ([4470a81](https://github.com/salesforcecli/plugin-dev/commit/4470a812c2219adf97f59bb116617d747ac78225))
+
 ### [0.0.7](https://github.com/salesforcecli/plugin-dev/compare/v0.0.6...v0.0.7) (2022-07-21)